+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/matcher/UrlMatcher.kt
@@ -15,20 +15,20 @@ import java.nio.charset.StandardCharsets.UTF_8
 import java.util.LinkedList
 
 /**
- * Provides functionality to process categorized URL black/white lists and match
+ * Provides functionality to process categorized URL block/safe lists and match
  * URLs against these lists.
  */
 class UrlMatcher {
     private val categories: MutableMap<String, Trie>
     internal val enabledCategories = HashSet<String>()
 
-    private val whiteList: WhiteList?
+    private val safelist: Safelist?
     private val previouslyMatched = HashSet<String>()
     private val previouslyUnmatched = HashSet<String>()
 
     constructor(patterns: Array<String>) {
         categories = HashMap()
-        whiteList = null
+        safelist = null
 
         val defaultCategory = Trie.createRootNode()
         patterns.forEach { defaultCategory.put(it.reverse()) }
@@ -40,9 +40,9 @@ class UrlMatcher {
         enabledCategories: Set<String>,
         supportedCategories: Set<String>,
         categoryMap: MutableMap<String, Trie>,
-        whiteList: WhiteList? = null
+        safelist: Safelist? = null
     ) {
-        this.whiteList = whiteList
+        this.safelist = safelist
         this.categories = categoryMap
 
         for ((key) in categoryMap) {
@@ -85,7 +85,7 @@ class UrlMatcher {
     }
 
     /**
-     * Checks if the given page URI is blacklisted for the given resource URI.
+     * Checks if the given page URI is blocklisted for the given resource URI.
      * Returns true if the site (page URI) is allowed to access
      * the resource URI, otherwise false.
      *
@@ -99,7 +99,7 @@ class UrlMatcher {
     }
 
     /**
-     * Checks if the given page URI is blacklisted for the given resource URI.
+     * Checks if the given page URI is blocklisted for the given resource URI.
      * Returns true if the site (page URI) is allowed to access
      * the resource URI, otherwise false.
      *
@@ -119,7 +119,7 @@ class UrlMatcher {
             return notMatchesFound
         }
 
-        if (whiteList?.contains(pageURI, resourceURI) == true) {
+        if (safelist?.contains(pageURI, resourceURI) == true) {
             return notMatchesFound
         }
 
@@ -170,54 +170,54 @@ class UrlMatcher {
          * Creates a new matcher instance for the provided URL lists.
          *
          * @deprecated Pass resources directly
-         * @param blackListFile resource ID to a JSON file containing the black list
-         * @param whiteListFile resource ID to a JSON file containing the white list
+         * @param blocklistFile resource ID to a JSON file containing the block list
+         * @param safelistFile resource ID to a JSON file containing the safe list
          */
         fun createMatcher(
             context: Context,
-            @RawRes blackListFile: Int,
-            @RawRes whiteListFile: Int,
+            @RawRes blocklistFile: Int,
+            @RawRes safelistFile: Int,
             enabledCategories: Set<String> = supportedCategories
         ): UrlMatcher =
-            createMatcher(context.resources, blackListFile, whiteListFile, enabledCategories)
+            createMatcher(context.resources, blocklistFile, safelistFile, enabledCategories)
 
         /**
          * Creates a new matcher instance for the provided URL lists.
          *
-         * @param blackListFile resource ID to a JSON file containing the black list
-         * @param whiteListFile resource ID to a JSON file containing the white list
+         * @param blocklistFile resource ID to a JSON file containing the block list
+         * @param safelistFile resource ID to a JSON file containing the safe list
          */
         fun createMatcher(
             resources: Resources,
-            @RawRes blackListFile: Int,
-            @RawRes whiteListFile: Int,
+            @RawRes blocklistFile: Int,
+            @RawRes safelistFile: Int,
             enabledCategories: Set<String> = supportedCategories
         ): UrlMatcher {
-            val blackListReader = InputStreamReader(resources.openRawResource(blackListFile), UTF_8)
-            val whiteListReader = InputStreamReader(resources.openRawResource(whiteListFile), UTF_8)
-            return createMatcher(blackListReader, whiteListReader, enabledCategories)
+            val blocklistReader = InputStreamReader(resources.openRawResource(blocklistFile), UTF_8)
+            val safelistReader = InputStreamReader(resources.openRawResource(safelistFile), UTF_8)
+            return createMatcher(blocklistReader, safelistReader, enabledCategories)
         }
 
         /**
          * Creates a new matcher instance for the provided URL lists.
          *
-         * @param black reader containing the black list
-         * @param white resource ID to a JSON file containing the white list
+         * @param block reader containing the block list
+         * @param safe resource ID to a JSON file containing the safe list
          */
         fun createMatcher(
-            black: Reader,
-            white: Reader,
+            block: Reader,
+            safe: Reader,
             enabledCategories: Set<String> = supportedCategories
         ): UrlMatcher {
             val categoryMap = HashMap<String, Trie>()
 
-            JsonReader(black).use {
+            JsonReader(block).use {
                 jsonReader -> loadCategories(jsonReader, categoryMap)
             }
 
-            var whiteList: WhiteList?
-            JsonReader(white).use { jsonReader -> whiteList = WhiteList.fromJson(jsonReader) }
-            return UrlMatcher(enabledCategories, supportedCategories, categoryMap, whiteList)
+            var safelist: Safelist?
+            JsonReader(safe).use { jsonReader -> safelist = Safelist.fromJson(jsonReader) }
+            return UrlMatcher(enabledCategories, supportedCategories, categoryMap, safelist)
         }
 
         /**